Improve hide button positioning
